pythonFEBE/requirements.txt
flask
openai
anthropic
python-dotenv
deepgram-sdk
httpx
verboselogs
pytube
moviepy